Switching from Lake/QL to CLPM
The current build process for this library is challenging to use if you
aren't using Roswell.  Lake also appears to be in a somewhat unhealthy
state.  This commit gets rid of the need for Lake and Roswell.  It
introduces a clpmfile and clpmfile.lock to manage dependencies.  It also
adds a small build script.  The README is also updated.