Add explicit (:USE) subform
Add explicit (:USE) subform to generated defpackages
Fixes PR pokepay#12 and makes it so that this system compiles cleanly on
Lispworks and CCL.

The original author of this commit was fiddlerwoaroof.  I should have
just merged his changes.  I chose copypasta for some unknown reason.
Here is the original work:
https://github.com/fiddlerwoaroof/aws-sdk-lisp/commit/96e77348b710a7da0ac78693fc9baf626d9be000